Fixed type error
